+++ src/components/users/user-warning.tsx
@@ -2,8 +2,14 @@ import Alert from "@/components/common/alert"
 
 const UserWarning = ({ type }: { type?: string }) => (
   <div className="user-warning">
-    {type === "no_alone_services" && (
+    {type === "alone_service" && (
+      <Alert message="Cet utilisateur n'est constitué que d'un seul service." />
+    )}
+    {type === "missing_services" && (
       <Alert message="Cet utilisateur n'a pas de compte Github ou Mattermost." />
+    )}
+    {type === "missing_departure" && (
+      <Alert message="Cet utilisateur n'a pas de date de départ." />
     )}
   </div>
 )