+++ src/utils/detect-warnings.ts
@@ -1,4 +1,7 @@
-const noAloneServices = (userEntry: Record<string, unknown>) => {
+import SERVICES from "@/utils/SERVICES"
+
+const containsMultipleServices = (userEntry: Record<string, unknown>) => {
+  // Services that should not be alone in an entry
   const FORBIDDEN_ALONE_SERVICES: ServiceName[] = [
     "sentry",
     "zammad",
@@ -8,15 +11,37 @@ const noAloneServices = (userEntry: Record<string, unknown>) => {
   ]
 
   const services = Object.keys(userEntry).filter(
-    (key) => key !== "id" && userEntry[key] !== null
+    (key) => (SERVICES as string[]).includes(key) && userEntry[key] !== null
   ) as ServiceName[]
   return services.length === 1 && FORBIDDEN_ALONE_SERVICES.includes(services[0])
-    ? "no_alone_services"
+    ? "alone_service"
     : null
 }
 
-// Insert a rule function here to active the rule
-const ACTIVE_RULES = [noAloneServices]
+const containsSpecificServices = (userEntry: Record<string, unknown>) => {
+  // We want to find at least one these services in each entry
+  const MANDATORY_SERVICES: ServiceName[] = ["github", "mattermost"]
+
+  const services = Object.keys(userEntry).filter(
+    (key) => (SERVICES as string[]).includes(key) && userEntry[key] !== null
+  ) as ServiceName[]
+  return services.some((element) => MANDATORY_SERVICES.includes(element))
+    ? null
+    : "missing_services"
+}
+
+const containsDepartureDate = (userEntry: Record<string, unknown>) => {
+  // We want to find a departure date in each entry
+  return userEntry.departure === null ? "missing_departure" : null
+}
+
+// Insert a rule function here to activate the rule
+// The retained warning will be the first found
+const ACTIVE_RULES = [
+  containsMultipleServices,
+  // containsDepartureDate,
+  containsSpecificServices,
+]
 
 export const detectWarnings = (
   userEntry: Record<string, Record<string, unknown>>